18-12-2023 Deposito

+++ tb/src/main/java/com/spring/tb/domain/model/Conta.java
@@ -20,7 +20,7 @@ public class Conta {
 
     private int numero;
 
-    private Double saldo;
+    private Float saldo;
 
     @JsonIgnore
     @OneToOne
